ci: :bug: common/Coil initial bug fix
