+++ src/pages/Common/Coil/Log/index.jsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
-import ProductionLog from './ProductionLog';
+import Production from './Production';
 import RMOrderAgainstLog from './RMOrderAgainstLog';
-import RMTransferLog from './RMTransfer';
-import SFGTransferLog from './SFGTransfer';
+import RMTransferLog from './RMTransferLog';
+import CoilToDyeing from './CoilToDyeing';
 
 export default function Index() {
 	useEffect(() => {
@@ -10,11 +10,11 @@ export default function Index() {
 	}, []);
 	return (
 		<div>
-			<SFGTransferLog />
-			<hr className='my-6 border-2 border-dashed border-secondary-content' />
 			<RMTransferLog />
 			<hr className='my-6 border-2 border-dashed border-secondary-content' />
-			<ProductionLog />
+			<CoilToDyeing />
+			<hr className='my-6 border-2 border-dashed border-secondary-content' />
+			<Production />
 			<hr className='my-6 border-2 border-dashed border-secondary-content' />
 			<RMOrderAgainstLog />
 		</div>
